Refactored to run agol_to_pg in flask instead of external process
Updated sql to handle errors vs failed to truncate because of foreign key.
Added vscode settings and launch to debug flask app inside vscode.
Update to use conn and cursor context for auto closing
Setup agol_to_pg to log to container logs as well as stream events
Added RESTART IDENTITY to sql in order to reset the main table sequence when truncated
Update OID to be required and not an existing field in the service to maintain OID values from the service
Moved API key validation to before request call so it can be auto re-used